refactor: reorganise files to decouple periods from standard reports
