+++ src/pages/standard-report/standard.report.conf.js
@@ -1,4 +1,5 @@
 import i18n from '@dhis2/d2-i18n'
+export { RELATIVE_PERIODS } from '../../utils/periods/relativePeriods'
 
 export const REPORTS_ENDPOINT = 'reports'
 export const REPORT_TEMPLATES_ENDPOINT = 'reportTemplate'
@@ -120,200 +121,3 @@ export const CACHE_STRATEGIES = [
         name: i18n.t('Cache for two weeks'),
     },
 ]
-
-export const relativePeriods = [
-    // Days
-    {
-        label: i18n.t('Days'),
-        periods: [
-            {
-                id: 'thisDay',
-                name: i18n.t('Reporting day'),
-            },
-            {
-                id: 'yesterday',
-                name: i18n.t('Yesterday'),
-            },
-            {
-                id: 'last3Days',
-                name: i18n.t('Last 3 days'),
-            },
-            {
-                id: 'last7Days',
-                name: i18n.t('Last 7 days'),
-            },
-            {
-                id: 'last14Days',
-                name: i18n.t('Last 14 days'),
-            },
-        ],
-    },
-    // Weeks
-    {
-        label: i18n.t('Weeks'),
-        periods: [
-            {
-                id: 'thisWeek',
-                name: i18n.t('Reporting week'),
-            },
-            {
-                id: 'lastWeek',
-                name: i18n.t('Last Week'),
-            },
-            {
-                id: 'last4Weeks',
-                name: i18n.t('Last 2 weeks'),
-            },
-            {
-                id: 'last12Weeks',
-                name: i18n.t('Last 12 weeks'),
-            },
-            {
-                id: 'last52Weeks',
-                name: i18n.t('Last 52 weeks'),
-            },
-            {
-                id: 'weeksThisYear',
-                name: i18n.t('Weeks this year'),
-            },
-        ],
-    },
-    // Months
-    {
-        label: i18n.t('Months'),
-        periods: [
-            {
-                id: 'thisMonth',
-                name: i18n.t('Reporting month'),
-            },
-            {
-                id: 'lastMonth',
-                name: i18n.t('Last month'),
-            },
-            {
-                id: 'last3Months',
-                name: i18n.t('Last 3 months'),
-            },
-            {
-                id: 'last6Months',
-                name: i18n.t('Last 6 months'),
-            },
-            {
-                id: 'last12Months',
-                name: i18n.t('Last 12 months'),
-            },
-            {
-                id: 'monthsThisYear',
-                name: i18n.t('Months this year'),
-            },
-        ],
-    },
-    // Bi-months
-    {
-        label: i18n.t('Bi-months'),
-        periods: [
-            {
-                id: 'thisBimonth',
-                name: i18n.t('Reporting bi-month'),
-            },
-            {
-                id: 'lastBimonth',
-                name: i18n.t('Last bi-month'),
-            },
-            {
-                id: 'last6BiMonths',
-                name: i18n.t('Last 6 bi-month'),
-            },
-            {
-                id: 'biMonthsThisYear',
-                name: i18n.t('Bi-months this year'),
-            },
-        ],
-    },
-    // Quarters
-    {
-        label: i18n.t('Quarters'),
-        periods: [
-            {
-                id: 'thisQuarter',
-                name: i18n.t('Reporting quarter'),
-            },
-            {
-                id: 'lastQuarter',
-                name: i18n.t('Last quarter'),
-            },
-            {
-                id: 'last4Quarters',
-                name: i18n.t('Last 4 quarters'),
-            },
-            {
-                id: 'quartersThisYear',
-                name: i18n.t('Quarters of reporting year'),
-            },
-        ],
-    },
-    // Six-Months
-    {
-        label: i18n.t('Six-Months'),
-        periods: [
-            {
-                id: 'thisSixMonth',
-                name: i18n.t('Reporting six-month'),
-            },
-            {
-                id: 'lastSixMonth',
-                name: i18n.t('Last six-month'),
-            },
-            {
-                id: 'last2SixMonths',
-                name: i18n.t('Last 2 six-months'),
-            },
-        ],
-    },
-    // Financial years
-    {
-        label: i18n.t('Financial Years'),
-        periods: [
-            {
-                id: 'thisFinancialYear',
-                name: i18n.t('Reporting financial year'),
-            },
-            {
-                id: 'lastFinancialYear',
-                name: i18n.t('Last financial year'),
-            },
-            {
-                id: 'last5FinancialYears',
-                name: i18n.t('Last 5 financial years'),
-            },
-        ],
-    },
-    // years
-    {
-        label: i18n.t('Years'),
-        periods: [
-            {
-                id: 'thisYear',
-                name: i18n.t('Reporting year'),
-            },
-            {
-                id: 'lastYear',
-                name: i18n.t('Last year'),
-            },
-            {
-                id: 'last5Years',
-                name: i18n.t('Last 5 years'),
-            },
-        ],
-    },
-]
-
-export const flattenedRelativePeriods = relativePeriods.reduce((acc, group) => {
-    group.periods.forEach(period => {
-        acc[period.id] = {
-            id: period.id,
-            displayName: period.name,
-        }
-    })
-    return acc
-}, {})
