+++ src/utils/periods/periodTypes.js
@@ -1,0 +1,4 @@
+import { fixedPeriodTranslations } from './fixedPeriods'
+
+export const isFixedPeriodType = periodType =>
+    !!fixedPeriodTranslations[periodType]
